requirements.txt
google-adk
google-genai
litellm
pyautogui
Pillow
python-dotenv